requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.3
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
cssselect==1.2.0
cssutils==2.6.0
frozenlist==1.3.3
greenlet==2.0.1
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.4
playwright==1.32.1
premailer==3.10.0
pyee==9.0.4
python-dotenv==1.0.0
requests==2.28.2
SQLAlchemy==2.0.9
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
yagmail==0.15.293
yarl==1.8.2
zipp==3.15.0
